Refactor: 중복/불필요 웹소켓 처리 파일 삭제 및 디렉토리 이동
