--- src/main/java/com/example/taste/common/websocket/CustomHandshakeHandler.java
@@ -1,26 +1,0 @@
-package com.example.taste.common.websocket;
-
-import java.security.Principal;
-import java.util.Map;
-
-import jakarta.servlet.http.HttpServletRequest;
-
-import org.springframework.http.server.ServerHttpRequest;
-import org.springframework.http.server.ServletServerHttpRequest;
-import org.springframework.stereotype.Component;
-import org.springframework.web.socket.WebSocketHandler;
-import org.springframework.web.socket.server.support.DefaultHandshakeHandler;
-
-@Component
-public class CustomHandshakeHandler extends DefaultHandshakeHandler {
-	@Override
-	protected Principal determineUser(ServerHttpRequest request,
-		WebSocketHandler wsHandler,
-		Map<String, Object> attributes) {
-		if (request instanceof ServletServerHttpRequest servletRequest) {
-			HttpServletRequest httpRequest = servletRequest.getServletRequest();
-			return httpRequest.getUserPrincipal(); // 세션에 저장된 인증자 반환
-		}
-		return null;
-	}
-}
